refactor(lb): added test for upstream pool
